blog posts are now root level for backwards compatibility
